Remove libevent; Add libuv
